refactor(test): rename drain-event-bus.ts to test_polling.ts (pv-qmm2)
The helpers waitFor and waitForStableCount poll for database state
settling, not event-bus drainage. Renaming the module so the filename
matches its actual contents. Snake_case matches the existing sibling
test_environment.ts.

Closes pv-qmm2. Follow-up from pv-6evr consistency audit.
